CDKProj2/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_lambda